split generic method mutability tests
